Introduce a wrong fix for ignored auto configuration
Putting the auto configuration class into the same package as the demo makes the test work but not because of auto configuration kicking in but due to normal class path scanning. This will lead to all sorts of funny errors, @ConditionalOn* will work for example but @AutoConfigureBefore and friends not.