Fix duplicate beans
Problem here was: The name of the beans. The Bean instantiated by the old method is named "bannerSupplier", like the method and that name together with the type was a unique match. Not so if we change the name to "defaultBannerSupplier". Then we must make sure that "defaultBannerSupplier" is only called when there is no banner supplier already.

I prefer explicit ordering in such cases, so i added @Order as well.